feat: TextInput component
